requirements.txt
torch~=2.0.1
numpy~=1.23.4
matplotlib~=3.4.3
pillow~=9.2.0
torchvision~=0.15.2
scikit-learn~=1.2.1
einops~=0.5.0
npy_append_array~=0.9.16
